fix(machodylib): restrict IsDyldSharedCacheLib to Darwin; stub returns false elsewhere
The dyld shared cache is macOS-specific. Split into shared_cache_darwin.go
(darwin implementation) and shared_cache.go (!darwin stub returning false),
consistent with the naming convention used elsewhere in the codebase.
